LPD-48032 Break DDM's MultipleSelect component in three: Main, MultipleSelect and MultipleSelectBase
